Removed format() method from SalesReporter
Allowing for multiple ways to return data formatting.

Created the SalesOutputInterface interface which requires the implementation of the output() method.

Created the HtmlOutput class which implements the SalesOutputInterface interface. Moved the format() method to HtmlOutput class, and changed the method name to output().

Added the $formatter parameter to SalesReporter::between() method. Returning the formatted output.

Upated routes/web.php to pass SalesRepository class to the SalesReporter class in the $report variable. Passing HtmlOutput as $formatter parameter to the SalesReporter::between() method.